cli(refactor): enable new cmds and cleanup old cmds (#2696)
