+++ cmd/cdc/main.go
@@ -14,10 +14,10 @@
 package main
 
 import (
-	"github.com/pingcap/ticdc/cmd"
+	"github.com/pingcap/ticdc/pkg/cmd"
 	_ "github.com/pingcap/tidb/types/parser_driver"
 )
 
 func main() {
-	cmd.Execute()
+	cmd.Run()
 }
