+++ pkg/cmd/cli/cli_capture.go
@@ -14,20 +14,68 @@
 package cli
 
 import (
+	cmdcontext "github.com/pingcap/ticdc/pkg/cmd/context"
 	"github.com/pingcap/ticdc/pkg/cmd/factory"
+	"github.com/pingcap/ticdc/pkg/cmd/util"
 	"github.com/spf13/cobra"
 )
 
+// captureOptions defines flags for the `cli capture` command.
+type captureOptions struct {
+	disableVersionCheck bool
+}
+
+// newCaptureOptions creates new captureOptions for the `cli capture` command.
+func newCaptureOptions() *captureOptions {
+	return &captureOptions{}
+}
+
+// addFlags receives a *cobra.Command reference and binds
+// flags related to template printing to it.
+func (o *captureOptions) addFlags(cmd *cobra.Command) {
+	if o == nil {
+		return
+	}
+
+	cmd.PersistentFlags().BoolVar(&o.disableVersionCheck, "disable-version-check", false, "Disable version check")
+	_ = cmd.PersistentFlags().MarkHidden("disable-version-check")
+}
+
+// run checks the TiCDC cluster version.
+func (o *captureOptions) run(f factory.Factory) error {
+	if o.disableVersionCheck {
+		return nil
+	}
+	ctx := cmdcontext.GetDefaultContext()
+	etcdClient, err := f.EtcdClient()
+	if err != nil {
+		return err
+	}
+
+	_, err = util.VerifyAndGetTiCDCClusterVersion(ctx, etcdClient)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
 // newCmdCapture creates the `cli capture` command.
 func newCmdCapture(f factory.Factory) *cobra.Command {
+	o := newCaptureOptions()
+
 	cmds := &cobra.Command{
 		Use:   "capture",
 		Short: "Manage capture (capture is a CDC server instance)",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return o.run(f)
+		},
 	}
 	cmds.AddCommand(
 		newCmdListCapture(f),
 		// TODO: add resign owner command
 	)
+
+	o.addFlags(cmds)
 
 	return cmds
 }
