+++ pkg/cmd/cli/cli_processor.go
@@ -14,19 +14,67 @@
 package cli
 
 import (
+	cmdcontext "github.com/pingcap/ticdc/pkg/cmd/context"
 	"github.com/pingcap/ticdc/pkg/cmd/factory"
+	"github.com/pingcap/ticdc/pkg/cmd/util"
 	"github.com/spf13/cobra"
 )
 
+// processorOptions defines flags for the `cli processor` command.
+type processorOptions struct {
+	disableVersionCheck bool
+}
+
+// newProcessorOptions creates new processorOptions for the `cli processor` command.
+func newProcessorOptions() *processorOptions {
+	return &processorOptions{}
+}
+
+// addFlags receives a *cobra.Command reference and binds
+// flags related to template printing to it.
+func (o *processorOptions) addFlags(cmd *cobra.Command) {
+	if o == nil {
+		return
+	}
+
+	cmd.PersistentFlags().BoolVar(&o.disableVersionCheck, "disable-version-check", false, "Disable version check")
+	_ = cmd.PersistentFlags().MarkHidden("disable-version-check")
+}
+
+// run checks the TiCDC cluster version.
+func (o *processorOptions) run(f factory.Factory) error {
+	if o.disableVersionCheck {
+		return nil
+	}
+	ctx := cmdcontext.GetDefaultContext()
+	etcdClient, err := f.EtcdClient()
+	if err != nil {
+		return err
+	}
+
+	_, err = util.VerifyAndGetTiCDCClusterVersion(ctx, etcdClient)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
 // newCmdProcessor creates the `cli processor` command.
 func newCmdProcessor(f factory.Factory) *cobra.Command {
+	o := newProcessorOptions()
+
 	command := &cobra.Command{
 		Use:   "processor",
 		Short: "Manage processor (processor is a sub replication task running on a specified capture)",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return o.run(f)
+		},
 	}
 
 	command.AddCommand(newCmdListProcessor(f))
 	command.AddCommand(newCmdQueryProcessor(f))
+
+	o.addFlags(command)
 
 	return command
 }
